stepping in effect

+++ scripts/logic/effects/property/amounts/AmountProvider.cs
@@ -7,14 +7,14 @@ namespace Lawfare.scripts.logic.effects.property.amounts;
 [GlobalClass]
 public abstract partial class AmountProvider : Resource, IAmountProvider
 {
-    [Export] private int Multiplier = 1;
+    [Export] private float Multiplier = 1;
 
     [Export] private int Offset;
 
     public int GetAmount(GameEvent gameEvent, ISubject subject)
     {
         var count = Count(gameEvent, subject);
-        return Offset + Multiplier * count;
+        return (int) (Offset + Multiplier * count);
     }
 
     protected abstract int Count(GameEvent gameEvent, ISubject subject);
